safari-534.53-branch/LayoutTests/platform/chromium-win/fast/workers/worker-location-expected.txt

Test WorkerLocation properties.

WorkerLocation: function WorkerLocation() { [native code] }
typeof location: object
location: file:<...>/fast/workers/resources/worker-common.js
location.href: file:<...>/fast/workers/resources/worker-common.js
location.protocol: file:
location.host: 
location.hostname: 
location.port: 
location.pathname: <...>/fast/workers/resources/worker-common.js
location.search: 
location.hash: 
DONE
